[C] Reorganize locale strings
