+++ client/src/config/app/locale/en/backend.json
@@ -1,0 +1,1 @@
+{}
