+++ client/src/config/app/locale/en/frontend.json
@@ -1,0 +1,1 @@
+{}
