+++ client/src/config/app/locale/en/global.json
@@ -1,0 +1,13 @@
+{
+  "locales": {
+    "en": "English",
+    "es": "Spanish",
+    "nl": "Dutch"
+  },
+  "glossary": {
+    "project_one": "project",
+    "project_other": "projects",
+    "project_collection_one": "project collection",
+    "project_collection_other": "project collections"
+  }
+}
