+++ client/src/config/app/locale/en/index.js
@@ -1,0 +1,12 @@
+/* eslint-disable import/extensions */
+import global from "./global.json";
+import backend from "./backend.json";
+import frontend from "./frontend.json";
+import reader from "./reader.json";
+
+export default {
+  global,
+  backend,
+  frontend,
+  reader
+};
