+++ client/src/config/app/locale/en/reader.json
@@ -1,0 +1,1 @@
+{}
