+++ client/src/config/app/locale/es.json
