+++ client/src/config/app/locale/index.js
@@ -1,10 +1,10 @@
 // Import a translation file for each available language
 /* eslint-disable import/extensions */
-import en from "./en.json";
+import en from "./en";
 import es from "./es.json";
 import nl from "./nl.json";
 
-export const localeData = {
+export default {
   en,
   es,
   nl
