+++ client/src/utils/i18n.js
@@ -1,6 +1,6 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
-import { localeData } from "./localeStrings";
+import translations from "config/app/locale";
 
 export const updateI18n = lang => {
   if (i18n.language !== lang) {
@@ -12,8 +12,10 @@ i18n
   .use(initReactI18next) // passes i18n down to react-i18next
   .init({
     // debug: true,
+    ns: ["global", "backend", "frontend", "reader"],
+    defaultNS: "global",
     lng: "en",
-    resources: localeData,
+    resources: translations,
     interpolation: {
       escapeValue: false
     },
